small agent compile fix, implementing testmode for random maze

+++ StarterCode/Agent.cpp
@@ -55,7 +55,7 @@ void Agent::solveMaze() {
 void Agent::printAndGuideSolve (std::vector<mcpp::Coordinate> completeVec) { 
     mcpp::MinecraftConnection mc;
 
-    for (int i = 0; i < completeVec.size(); i++) {
+    for (unsigned long int i = 0; i < completeVec.size(); i++) {
         int j = i + 1;
         std::cout << "Step[" << j << "]: (" << completeVec[i].x << ", " << completeVec[i].y << ", " << completeVec[i].z << ")"  << std::endl;
         mc.setBlock(completeVec[i], mcpp::Blocks::LIME_CARPET);
